requirements.txt
flask==2.3.2
PyJWT==2.8.0
psycopg2
cryptography==42.0.5
requests
python-dotenv
py-healthcheck
SQLAlchemy